CTFS-formated data preparation

Main routine to format and correct CTFS-formated data

a wrapper to get smoothed predictions of AGB fluxes using a loess function (library 'locfit')

loess.function(x, var, range)

Arguments

x
a data.table

var
the name of the variable to be smoothed again intial AGB

range
the range of initial AGB to be used for prediction (i.e. 5th and 95th percentiles of the whole distribution)

site
provide the full name of your site (in lower case) i.e. 'barro colorado island'

stem
TRUE or FALSE, are you using the stem data (stem=TRUE) or tree data (i.e. called 'full')?

taper.correction
TRUE or FALSE, are you willing to apply Cushman et al (2014) taper correction?

fill_missing
TRUE or FALSE, are you willing to extrapolate missing DBH from surrounding DBH?

palm
TRUE or FALSE, if TRUE, biomass of palm trees is computed through a specific allometric model (Goodman et al. 2013)

strangler
TRUE or FALSE, if TRUE, strangler figs tree are flagged (upon a list to published soon)

maxrel
a numeric value setting the threshold over which relative productivity is assumed to be too high (usually set at 20 percents)

draw.graph
TRUE or FALSE, if TRUE, draw graph for trees with 'erroneous' DBH measures/productivity

output.errors
TRUE or FALSE, if TRUE, creat a CSV file with all trees with erroneous DBH measures/productivity

DATA_path
allows to provide a different path where the data are located

exclude.interval
a vector (i.e. c(1,2)) indicating if a set of census intervals must be discarded from computation due for instance to a change in protocol of measurment

Value

a data.table (data.frame) with all relevant variables.

a smoothed prediction of the variable of interest